@@ -7,16 +7,16 @@
         }
     },
     "require": {
-        "php": "^7.4 || ^8.0",
+        "php": ">=7.4",
         "paysera/lib-serializer": "^3.0",
         "doctrine/orm": "^2.0",
-        "symfony/config": "^3.0 || ^4.0",
-        "symfony/dependency-injection": "^3.0 || ^4.0",
-        "symfony/validator": "^3.0 || ^4.0",
-        "symfony/security-core": "^3.0 || ^4.0",
-        "symfony/routing": "^3.0 || ^4.0",
-        "symfony/http-foundation": "^3.0 || ^4.0",
-        "symfony/http-kernel": "^3.0 || ^4.0",
+        "symfony/config": "^4.0 || ^5.0",
+        "symfony/dependency-injection": "^4.0 || ^5.0",
+        "symfony/validator": "^4.0 || ^5.0",
+        "symfony/security-core": "^4.0 || ^5.0",
+        "symfony/routing": "^4.0 || ^5.0",
+        "symfony/http-foundation": "^4.0 || ^5.0",
+        "symfony/http-kernel": "^4.0 || ^5.0",
         "doctrine/persistence": "^3.2"
     },
     "config": {
@@ -25,5 +25,15 @@
     "require-dev": {
         "phpunit/phpunit": "^8.0",
         "mockery/mockery": "^1.2"
+    },
+    "scripts": {
+        "test": [
+            "@test:phpunit"
+        ],
+        "test:phpunit": "bin/phpunit"
+    },
+    "scripts-descriptions": {
+        "test": "Runs all tests",
+        "test:phpunit": "Runs PHPUnit tests"
     }
 }
